refactor: migrate form components to a new structure and update imports
- Renamed and exported ButtonGroup as a standalone component.
- Removed deprecated FormErrorMessage and FormLegend components.
- Introduced new form components: Caption, Checkbox, ControlGroup, Fieldset, Input, Label, Legend, Message, Option, Radio, Select, Separator, Textarea.
- Updated imports across various components to use the new form structure.
- Adjusted styling and class names for consistency.
- Refactored error handling to utilize the new Message component.
