+++ packages/ui/src/components/filters/filter-trigger.tsx
@@ -1,6 +1,6 @@
 import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from '@components/dropdown-menu'
+import { Input } from '@components/form'
 import { Icon } from '@components/icon'
-import { Input } from '@components/input'
 import { TFunction } from 'i18next'
 
 import { FilterOption, SortOption } from './types'
