+++ packages/ui/src/components/filters/filters-bar/filter-variants/number.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react'
 
+import { Input } from '@components/form'
 import { Icon } from '@components/icon'
-import { Input } from '@components/input'
 
 import { FilterValue } from '../../types'
 import { UseFiltersReturn } from '../../use-filters'
