+++ packages/ui/src/components/filters/filters-bar/filter-variants/text.tsx
@@ -1,8 +1,8 @@
 import { useState } from 'react'
 
 import { DropdownMenuItem } from '@components/dropdown-menu'
+import { Input } from '@components/form'
 import { Icon } from '@components/icon'
-import { Input } from '@components/input'
 
 import { FilterValue } from '../../types'
 import { UseFiltersReturn } from '../../use-filters'
