+++ packages/ui/src/components/form/separator.tsx
@@ -6,6 +6,21 @@ interface FormSeparatorProps {
   className?: string
 }
 
+/**
+ * A horizontal separator component for forms
+ * @component
+ * @param {FormSeparatorProps} props - The component props
+ * @returns {JSX.Element} A horizontal separator element
+ * @example
+ * // Basic usage
+ * <FormSeparator />
+ *
+ * // With dashed style
+ * <FormSeparator dashed />
+ *
+ * // With custom className
+ * <FormSeparator className="my-4" />
+ */
 export function FormSeparator({ dashed, dotted, className }: FormSeparatorProps) {
   return (
     <div
