+++ packages/ui/src/components/index.ts
@@ -1,7 +1,6 @@
 export * from './alert'
 export * from './icon'
 export * from './button'
-export * from './input'
 export * from './more-submenu'
 export * from './settings-menu'
 export * from './sheet'
@@ -17,34 +16,24 @@ export * from './alert-dialog'
 export * from './popover'
 export * from './avatar'
 export * from './navbar'
-export * from './label'
-export * from './select'
 export * from './theme-selector'
 export * from './theme-selector/types'
 export * from './no-data'
 export * from './tabs'
 export * from './command'
 export * from './search-files'
-export * from './radio-group'
-export * from './checkbox'
 export * from './toast'
 export * from './styled-link'
-export * from './fieldset'
-export * from './control-group'
-export * from './form-error-message'
 export * from './skeleton'
 export * from './table'
 export * from './commit-copy-actions'
-export * from './textarea'
 export * from './accordion'
 export * from './breadcrumb'
 export * from './skeleton-list'
 export * from './dialog'
 export * from './path-breadcrumbs'
-export * from './card'
-export * from './input-otp'
-export * from './control-group'
-export * from './form-error-message'
+export * from './form'
+export * from './button-group'
 
 export * as ShaBadge from './sha-badge'
 export * as ListActions from './list-actions'
@@ -52,7 +41,6 @@ export * as ListPagination from './list-pagination'
 export * as SearchBox from './search-box'
 export * as NavbarProjectChooser from './navbar-project-chooser'
 export * as StackedList from './stacked-list'
-export * as ButtonGroup from './button-group'
 export * as FileExplorer from './file-explorer'
 
 export type * from './navbar/types'
