+++ packages/ui/src/components/manage-navigation/index.tsx
@@ -127,11 +127,11 @@ export const ManageNavigation = ({
         </AlertDialogHeader>
         <ScrollArea className="-mx-5 -mb-5 mt-1">
           <div className="px-5">
-            <Text className="inline-block leading-none text-foreground-7" size={1}>
+            <Text className="text-foreground-7 inline-block leading-none" size={1}>
               Pinned
             </Text>
             {!currentPinnedItems.length ? (
-              <Text className="mt-3 block text-foreground-5" size={1}>
+              <Text className="text-foreground-5 mt-3 block" size={1}>
                 No pinned items
               </Text>
             ) : (
@@ -152,7 +152,7 @@ export const ManageNavigation = ({
                               <Text className="w-full text-left text-[inherit]">{item.title}</Text>
                             </Button>
                             <Button
-                              className="absolute right-1 top-0.5 z-20 text-icons-4 hover:text-icons-2"
+                              className="text-icons-4 hover:text-icons-2 absolute right-1 top-0.5 z-20"
                               size="sm_icon"
                               variant="custom"
                               onClick={() => removeFromPinnedItems(item)}
@@ -170,7 +170,7 @@ export const ManageNavigation = ({
             {currentFilteredRecentItems.length > 0 && (
               <>
                 <div className="mt-4 flex items-center justify-between">
-                  <Text className="inline-block leading-none text-foreground-7" size={1}>
+                  <Text className="text-foreground-7 inline-block leading-none" size={1}>
                     Recent
                   </Text>
                   <Button className="-mr-1.5" variant="link_accent" size="xs" onClick={handleClearRecent}>
@@ -182,10 +182,10 @@ export const ManageNavigation = ({
                     <li className="relative flex h-8 items-center" key={`recent-${item.id}-${index}`}>
                       <div className="flex w-full grow items-center gap-x-2.5">
                         <Icon className="text-icons-4" name="clock-icon" size={14} />
-                        <Text className="w-full text-left text-foreground-8">{item.title}</Text>
+                        <Text className="text-foreground-8 w-full text-left">{item.title}</Text>
                       </div>
                       <Button
-                        className="absolute -right-2 top-0.5 text-icons-4 hover:text-icons-2"
+                        className="text-icons-4 hover:text-icons-2 absolute -right-2 top-0.5"
                         size="sm_icon"
                         variant="custom"
                         onClick={() => addToPinnedItems(item)}
@@ -200,7 +200,7 @@ export const ManageNavigation = ({
           </div>
         </ScrollArea>
         <AlertDialogFooter>
-          <ButtonGroup.Root>
+          <ButtonGroup>
             {!submitted ? (
               <>
                 <Button variant="outline" onClick={handleCancel} disabled={isSubmitting}>
@@ -223,7 +223,7 @@ export const ManageNavigation = ({
                 <Icon name="tick" size={14} />
               </Button>
             )}
-          </ButtonGroup.Root>
+          </ButtonGroup>
         </AlertDialogFooter>
       </AlertDialogContent>
     </AlertDialog>
